- create core/models.py with 9 request/response models - Move AnalyzeRequest, ChatRequest, OpenShiftAnalyzeRequest, etc. from api/mcp.py - Update imports in api/mcp.py to use core.models - Remove duplicate model definitions from API layer Signed-off-by: Abdul Hameed <ahameed@redhat.com>
